+++ CHANGELOG.md
@@ -202,6 +202,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - A `send` method to `CommandError` which can be overridden to customise its response behaviour.
 - Support for other create message arguments to `CommandError`.
+- Support for toggleable case-insensitive message command name handling.
 
 ### Changed
 - Renamed `CommandError.message` to `.content`.